:recycle: Restructure files
