--- packages/lisk-transactions/src/transactions/helpers/get_id.ts
@@ -1,30 +1,0 @@
-/*
- * Copyright © 2018 Lisk Foundation
- *
- * See the LICENSE file at the top-level directory of this distribution
- * for licensing information.
- *
- * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
- * no part of this software, including this file, may be copied, modified,
- * propagated, or distributed except according to the terms contained in the
- * LICENSE file.
- *
- * Removal or modification of this copyright notice is prohibited.
- *
- */
-import * as cryptography from '@liskhq/lisk-cryptography';
-import { TransactionJSON } from '../../transaction_types';
-import { getTransactionBytes } from '../../utils';
-
-export const getId = (transaction: TransactionJSON): string => {
-	const transactionBytes = getTransactionBytes(transaction);
-	const transactionHash = cryptography.hash(transactionBytes);
-	const bufferFromFirstEntriesReversed = cryptography.getFirstEightBytesReversed(
-		transactionHash,
-	);
-	const transactionId = cryptography.bufferToBigNumberString(
-		bufferFromFirstEntriesReversed,
-	);
-
-	return transactionId;
-};
