--- packages/lisk-transactions/src/transactions/helpers/index.ts
@@ -1,29 +1,0 @@
-/*
- * Copyright © 2018 Lisk Foundation
- *
- * See the LICENSE file at the top-level directory of this distribution
- * for licensing information.
- *
- * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
- * no part of this software, including this file, may be copied, modified,
- * propagated, or distributed except according to the terms contained in the
- * LICENSE file.
- *
- * Removal or modification of this copyright notice is prohibited.
- *
- */
-import { checkTypes } from './check_types';
-import { getBytes } from './get_bytes';
-import { getId } from './get_id';
-import { verifyBalance } from './verify_balance';
-import { verifyMultisignatures } from './verify_multisignatures';
-import { verifySignature } from './verify_signature';
-
-export {
-	checkTypes,
-	getBytes,
-	getId,
-	verifyBalance,
-	verifyMultisignatures,
-	verifySignature,
-};
