+++ packages/lisk-transactions/src/utils/get_transaction_id.ts
@@ -16,6 +16,19 @@ import * as cryptography from '@liskhq/lisk-cryptography';
 import { TransactionJSON } from '../transaction_types';
 import { getTransactionBytes } from './get_transaction_bytes';
 
+export const getId = (transactionBytes: Buffer): string => {
+	const transactionHash = cryptography.hash(transactionBytes);
+	const bufferFromFirstEntriesReversed = cryptography.getFirstEightBytesReversed(
+		transactionHash,
+	);
+	const transactionId = cryptography.bufferToBigNumberString(
+		bufferFromFirstEntriesReversed,
+	);
+
+	return transactionId;
+};
+
+// FIXME: Deprecated
 export const getTransactionId = (transaction: TransactionJSON): string => {
 	const transactionBytes = getTransactionBytes(transaction);
 	const transactionHash = cryptography.hash(transactionBytes);
