+++ packages/lisk-transactions/test/helpers/test_transaction_class.ts
@@ -1,9 +1,9 @@
+import * as cryptography from '@liskhq/lisk-cryptography';
 import {
 	BaseTransaction,
 	TransactionResponse,
 } from '../../src/transactions/base';
 import { TransactionJSON, Status } from '../../src/transaction_types';
-import { getBytes } from '../../src/transactions/helpers';
 
 export class TestTransaction extends BaseTransaction {
 	public containsUniqueData() {
@@ -14,9 +14,21 @@ export class TestTransaction extends BaseTransaction {
 		return {};
 	}
 
-	public getBytes() {
-		const transactionBytes = getBytes(this.toJSON());
+	public getAssetBytes() {
+		return Buffer.alloc(0);
+	}
 
+	public getBytes() {
+		const transactionBytes = Buffer.concat([
+			this.getBasicBytes(),
+			this.getAssetBytes(),
+			this.signature
+				? cryptography.hexToBuffer(this.signature)
+				: Buffer.alloc(0),
+			this.signSignature
+				? cryptography.hexToBuffer(this.signSignature)
+				: Buffer.alloc(0),
+		]);
 		return transactionBytes;
 	}
